Fix interactive buffers
