+++ src/EditorFeatures/Core.Wpf/Interactive/InteractiveEvaluator.cs
@@ -26,6 +26,7 @@ using Roslyn.Utilities;
 namespace Microsoft.CodeAnalysis.Interactive
 {
     using InteractiveHost::Microsoft.CodeAnalysis.Interactive;
+    using Microsoft.VisualStudio.Text.Editor;
 
     // TODO: Rename to InteractiveEvaluator https://github.com/dotnet/roslyn/issues/6441
     // The code is not specific to C#, but Interactive Window has hardcoded "CSharpInteractiveEvaluator" name.
@@ -71,6 +72,7 @@ namespace Microsoft.CodeAnalysis.Interactive
             IInteractiveWindowCommandsFactory commandsFactory,
             ImmutableArray<IInteractiveWindowCommand> commands,
             ITextDocumentFactoryService textDocumentFactoryService,
+            IEditorOptionsFactoryService editorOptionsFactory,
             InteractiveEvaluatorLanguageInfoProvider languageInfo,
             string initialWorkingDirectory)
         {
@@ -85,7 +87,16 @@ namespace Microsoft.CodeAnalysis.Interactive
 
             _workspace = new InteractiveWindowWorkspace(hostServices, globalOptions);
 
-            _session = new InteractiveSession(_workspace, threadingContext, listener, textDocumentFactoryService, languageInfo, initialWorkingDirectory);
+            _session = new InteractiveSession(
+                _workspace,
+                threadingContext,
+                listener,
+                textDocumentFactoryService,
+                editorOptionsFactory,
+                globalOptions,
+                languageInfo,
+                initialWorkingDirectory);
+
             _session.Host.ProcessInitialized += ProcessInitialized;
         }
