+++ src/VisualStudio/CSharp/Impl/Interactive/CSharpVsInteractiveWindowProvider.cs
@@ -18,6 +18,7 @@ using Microsoft.VisualStudio.LanguageServices.Interactive;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Classification;
+using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Utilities;
 using LanguageServiceGuids = Microsoft.VisualStudio.LanguageServices.Guids;
 
@@ -30,6 +31,7 @@ namespace Microsoft.VisualStudio.LanguageServices.CSharp.Interactive
         private readonly IAsynchronousOperationListener _listener;
         private readonly IGlobalOptionService _globalOptions;
         private readonly ITextDocumentFactoryService _textDocumentFactoryService;
+        private readonly IEditorOptionsFactoryService _editorOptionsFactory;
 
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
@@ -44,6 +46,7 @@ namespace Microsoft.VisualStudio.LanguageServices.CSharp.Interactive
             [ImportMany] IInteractiveWindowCommand[] commands,
             IGlobalOptionService globalOptions,
             ITextDocumentFactoryService textDocumentFactoryService,
+            IEditorOptionsFactoryService editorOptionsFactory,
             VisualStudioWorkspace workspace)
             : base(serviceProvider, interactiveWindowFactory, classifierAggregator, contentTypeRegistry, commandsFactory, commands, workspace)
         {
@@ -51,6 +54,7 @@ namespace Microsoft.VisualStudio.LanguageServices.CSharp.Interactive
             _listener = listenerProvider.GetListener(FeatureAttribute.InteractiveEvaluator);
             _globalOptions = globalOptions;
             _textDocumentFactoryService = textDocumentFactoryService;
+            _editorOptionsFactory = editorOptionsFactory;
         }
 
         protected override Guid LanguageServiceGuid => LanguageServiceGuids.CSharpLanguageServiceId;
@@ -78,6 +82,7 @@ namespace Microsoft.VisualStudio.LanguageServices.CSharp.Interactive
                 CommandsFactory,
                 Commands,
                 _textDocumentFactoryService,
+                _editorOptionsFactory,
                 CSharpInteractiveEvaluatorLanguageInfoProvider.Instance,
                 Environment.GetFolderPath(Environment.SpecialFolder.UserProfile));
         }
